Classic DOOM - MBF21 - Implement most new Thing flags (MF2)

+++ doomclassic/doom/info.h
@@ -1384,7 +1384,7 @@ typedef struct
     int infightingGroup;
     int projectileGroup;
     int splashGroup;
-    int flags2;
+    uint64 flags2;
 
 } mobjinfo_t;
